% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDAU.R
\name{addGroupingScheme}
\alias{addGroupingScheme}
\title{Add a custom grouping scheme.}
\usage{
addGroupingScheme(color = vector("character"), symbol = vector("character"),
  group = list())
}
\arguments{
\item{color}{A vector of character, with length match the number of desired 
groups and names as the same as the groups' names. This vector specifies the 
different colors for visualizing the different groups of amino acids.}

\item{symbol}{A vector of character, with length match the number of desired 
groups and names as the same as the groups' names. This vector specifies the 
different symbols for visualizing the different groups of amino acids.}

\item{group}{A list with names as the same as the groups' names.}
}
\value{
Add the custom grouping scheme to the cached environment.
}
\description{
Add a custom grouping scheme for grouping amino acids as desired.
}
\examples{
## Add a grouping scheme based on the BLOSUM50 level 3 
color = c(LVIMC = "#33FF00", AGSTP = "#CCFF00",
         FYW = '#00FF66', EDNQKRH = "#FF0066")
symbol = c(LVIMC = "L", AGSTP = "A", FYW = "F", EDNQKRH = "E")
group = list(
   LVIMC = c("L", "V", "I", "M", "C"), 
   AGSTP = c("A", "G", "S", "T", "P"),
   FYW = c("F", "Y", "W"),
   EDNQKRH = c("E", "D", "N", "Q", "K", "R", "H"))
addGroupingScheme(color = color, symbol = symbol, group = group) 

}
