% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachedEnv.R
\name{availableSchemes}
\alias{availableSchemes}
\title{Get all predefined coloring and grouping schemes}
\usage{
availableSchemes()
}
\value{
A vector of names of predefined coloring and grouping schemes stored in the environment `cacheEnv`.
}
\description{
List all predefined coloring and grouping schemes stored in the environmetn `cacheEnv`
}
\author{
Haibo Liu
}
\keyword{misc}
