% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBackgroundModel.R
\name{buildBackgroundModel}
\alias{buildBackgroundModel}
\title{Build background models for DAU tests}
\usage{
buildBackgroundModel(dagPeptides, background = c("wholeProteome", "inputSet",
  "nonInputSet"), model = c("any", "anchored"), targetPosition = c("any",
  "Nterminus", "Cterminus"), uniqueSeq = TRUE, numSubsamples = 300L,
  rand.seed = 1, replacement = FALSE, testType = c("ztest", "fisher"),
  proteome)
}
\arguments{
\item{dagPeptides}{An object of \code{\link{dagPeptides}} class containing 
peptide sequences as the input set.}

\item{background}{A character vector with defaults: "wholeProteome" and "inputSet", 
"nonInputSet", indicating what set of peptide sequences should be considered to 
generate a background model.}

\item{model}{A character vector with defaults: "any" and "anchored", indicating 
whether an anchoring position should be applied to generate a background model.}

\item{targetPosition}{A character vector with defaults: "any", "Nterminus" and 
"Cterminus", indicating whether which part of protein sequences of choice should
be used to generate a background model.}

\item{uniqueSeq}{A logical vector indicating whether only unique peptide sequences
are included in a background model.}

\item{numSubsamples}{An integer, the number of random sampling.}

\item{rand.seed}{An integer, the seed used to perform random sampling}

\item{replacement}{A logical vector of length 1, indicating whether replacement 
is allowed for random sampling.}

\item{testType}{A character vector of length 1. Available options are "ztest" 
and "fisher".}
}
\value{
An object of \code{\link{dagBackground}} class.
}
\description{
A method used to build background models for testing differential amino acid usage
}
\examples{
dat <- unlist(read.delim(system.file(
                                   "extdata", "grB.txt", package = "dagLogo"),
                         header = F, as.is = TRUE))
head(dat)
##prepare proteome from a fasta file
proteome <- prepareProteome(fasta = system.file("extdata",
                                                "HUMAN.fasta",
                                                package = "dagLogo"), 
                            species = "Homo sapiens")
##prepare object of dagPeptides
##prepare an object of dagPeptides
seq <- formatSequence(seq = dat, proteome = proteome, upstreamOffset = 14,
                     downstreamOffset = 15)
bg_fisher <- buildBackgroundModel(seq, background = "wholeProteome", 
                                  proteome = proteome, testType = "fisher")
bg_ztest <- buildBackgroundModel(seq, background = "wholeProteome",
                                   proteome = proteome, testType = "ztest")
}
\author{
Jianhong Ou, Haibo Liu
}
