% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPeptides.R
\name{cleanPeptides}
\alias{cleanPeptides}
\title{clean up peptides}
\usage{
cleanPeptides(dat, anchors)
}
\arguments{
\item{dat}{input data. The input dat contains two columns symbol and peptides.
the anchor AA must be in lower case.}

\item{anchors}{a vector of character, anchor Amino Acid, must be lower case.}
}
\value{
an data.frame with columns: `symbol`, `peptides` and `anchor`
}
\description{
clean the input file. The function removes peptides without any anchors, 
adds peptide for each additional anchor in each peptide, and allows multiple anchors as input.
}
\examples{
dat <- read.csv(system.file("extdata", "peptides2filter.csv", package="dagLogo"))
dat
dat.new <- cleanPeptides(dat, anchors = c("s", "t"))
}
\author{
Jianhong Ou, Julie Zhu
}
\keyword{misc}
