\name{dagBackground-class}
\docType{class}
\alias{dagBackground-class}

\title{Class \code{"dagBackground"}}
\description{
  An object of class \code{"dagBackground"} represents background model.
}

\section{Objects from the Class}{
  Objects can be created by calls of the form 
  \code{new("dagBackground", background, permutationSize)}.
}
\section{Slots}{
  \describe{
    \item{\code{background}}{Object of class \code{"list"} records the background model}
    \item{\code{permutationSize}}{code{"integer"} permutation size of background}
  }
}
\keyword{classes}
