% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagLogoPackage.R
\docType{package}
\name{dagLogo-package}
\alias{dagLogo}
\alias{dagLogo-package}
\title{Perform differential analysis of amino acid usage.}
\description{
dagLogo provides differential analysis of grouped/ungrouped amino acid usage
between an input set of aligned peptide sequences and a background set of 
aligned peptide sequences which can be generated in different ways. Results 
of Fisher's exact test and/or Z-test are visualized using a heatmap or DAG Logo.
}
