% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagLogo.R
\name{dagLogo}
\alias{dagLogo}
\title{Create sequence logo.}
\usage{
dagLogo(testDAUresults, type = c("diff", "zscore"), pvalueCutoff = 0.05,
  groupingSymbol = getGroupingSymbol(testDAUresults@group),
  font = "Helvetica-Bold", fontface = "bold", fontsize = 5,
  legend = FALSE, labelRelativeToAnchor = FALSE,
  gscmd = Sys.getenv("R_GSCMD"), labels = NULL)
}
\arguments{
\item{testDAUresults}{An object of \code{testDAUresults} (results of testing 
differential amino acid uasage).}

\item{type}{A character vector of length 1. Type of statistics to display 
on y-axis, available choices are "diff" or "zscore".}

\item{pvalueCutoff}{A numeric vector of length 1. A cutoff of p-value.}

\item{groupingSymbol}{A named character vector, with 3-letter symbols of amino acids
as names and single-letter symbols as values.}

\item{font}{A character vector of length 1. Font type for displaying sequence
Logo.}

\item{fontface}{An integer, fontface of text for axis annotation and legends.}

\item{fontsize}{An integer, fontsize of text for axis annotation and legends.}

\item{legend}{A logical vector of length 1, indicating whether to show the 
legend.}

\item{labelRelativeToAnchor}{A logical vector of length 1, indicating whether
x-axis label should be adjusted relative to the anchoring position.}

\item{gscmd}{A character vector of length 1, indicating the file path to 
Ghostscript. This should be defined as a system variable or provide directly.}

\item{labels}{A character vector, x-axis labels.}
}
\value{
A sequence Logo is plotted without returned values.
}
\description{
Create sequence logo for visualizing results of testing differential usage 
of amino acids.
}
\examples{
data('seq.example') 
data('proteome.example')
bg <- buildBackgroundModel(seq.example, proteome=proteome.example, 
                           numSubsamples=10, testType = "ztest")
t0 <- testDAU(seq.example, bg)
t1 <- testDAU(dagPeptides = seq.example, dagBackground = bg, 
              groupingScheme = "classic")
t2 <- testDAU(dagPeptides = seq.example, dagBackground = bg, 
             groupingScheme = "charge")
t3 <- testDAU(dagPeptides = seq.example, dagBackground = bg, 
             groupingScheme = "chemistry")
t4 <- testDAU(dagPeptides = seq.example, dagBackground = bg, 
              groupingScheme = "hydrophobicity")
dagLogo(t0)
dagLogo(t1, groupingSymbol = getGroupingSymbol(t1@group))
dagLogo(t2, groupingSymbol = getGroupingSymbol(t2@group))
dagLogo(t3, groupingSymbol = getGroupingSymbol(t3@group))
dagLogo(t4, groupingSymbol = getGroupingSymbol(t4@group))
}
\author{
Jianhong Ou, Haibo Liu
}
