% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclass.R
\docType{class}
\name{dagPeptides-class}
\alias{dagPeptides-class}
\title{Class dagPeptides.
An S4 class to represent formatted, aligned peptides for DAGlogo analysis.}
\description{
Class dagPeptides.
An S4 class to represent formatted, aligned peptides for DAGlogo analysis.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with columns: IDs, anchorAA, anchorPos, peptide and anchor.}

\item{\code{peptides}}{A matrix of character, each element is a single-character symbol
for a amino acid.}

\item{\code{upstreamOffset}}{An integer in the interval (0, 20): the upstream offset
relative to the anchoring position.}

\item{\code{downstreamOffset}}{An integer in the interval (0, 20): the downstream 
offset relative to the anchoring position.}

\item{\code{type}}{A character vector of length 1. Available options :"Uniprot", and 
"fasta".}
}}

\author{
Jianhong Ou
}
