% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli.proteome}
\alias{ecoli.proteome}
\title{An object of \code{\link{Proteome}} representing the E. coli proteome.}
\format{A object of \code{\link{Proteome}} Class}
\source{
\url{http://www.uniprot.org/}
}
\usage{
ecoli.proteome
}
\description{
A dataset containing the E. coli proteome. The variables are as follows:
}
\details{
\itemize{
  \item proteome. A data frame containing ENTREZ_GENE IDs, protein SEQUENCES, 
                  and UniProtKB IDs and length of the sequences
  \item type. The data source: Uniprot
  \item species. Escherichia coli
}
}
\keyword{datasets}
