% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSequence.R
\name{formatSequence}
\alias{formatSequence}
\title{Format already aligned peptide sequences.}
\usage{
formatSequence(seq, proteome, upstreamOffset, downstreamOffset)
}
\arguments{
\item{seq}{A vector of aligned peptide sequences of the same length}

\item{proteome}{An object of \code{Proteome} Class.}

\item{upstreamOffset}{An integer in the interval (0, 20): the upstream offset
relative to the anchoring position.}

\item{downstreamOffset}{An integer in the interval (0, 20): the downstream 
offset relative to the anchoring position.}
}
\value{
An object of dagPeptides Class
}
\description{
Convert already aligned peptide sequences into an object of \code{dagPeptides} 
class.
}
\examples{
## You already have the aligned peptides sequences in hand. Then you can use 
the formatSequence function to prepare an object of dagPeptides. Befor doing 
that, you need prepare a Proteome object by the prepareProteome function.

dat <- unlist(read.delim(system.file(
                                   "extdata", "grB.txt", package = "dagLogo"),
                         header = F, as.is = TRUE))
head(dat)
##prepare proteome from a fasta file
proteome <- prepareProteome(fasta = system.file("extdata",
                                                "HUMAN.fasta",
                                                package = "dagLogo"), 
                            species = "Homo sapiens")
##prepare object of dagPeptides
##prepare an object of dagPeptides
seq <- formatSequence(seq = dat, proteome = proteome, upstreamOffset = 14,
                     downstreamOffset = 15)

}
\author{
Jianhong Ou
}
