% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagLogoPlot.R
\name{getData}
\alias{getData}
\title{Get the data for visualization.}
\usage{
getData(type, testDAUresults)
}
\arguments{
\item{type}{A character vector of length 1, the type of metrics to display 
on y-axis. The available options are "diff" and "statistics", which are 
differences in amino acide usage at each position between the inputSet and 
the backgroundSet, and the Z-scores or odds ratios when Z-test or Fisher's 
exact test is performed to test the differential uasge of amino acid at each 
position between the two sets.}

\item{testDAUresults}{An object of \code{\link{testDAUresults}} Class (results
of testing differential amino acid usage).}
}
\value{
A list containing the following compoenets:
label A character vector of length 1. The type of data for visualization.
dat   A matrix of numeric data for visualization.
}
\description{
A helper function to get the data and it lable for visualization.
}
\examples{

}
\author{
Haibo Liu
}
\keyword{internal}
