% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagLogo.R
\name{getGroupingSymbol}
\alias{getGroupingSymbol}
\title{Get character symbols for grouped amino acids}
\usage{
getGroupingSymbol(groupingScheme = c("no", "classic", "charge", "chemistry",
  "hydrophobicity", "BLOSM50_L1", "BLOSM50_LL2", "BLOSM50_L3", "BLOSM50_L4",
  "BLOSM50_L5", "BLOSM50_L6", "BLOSM50_L7", "BLOSM50_L8",
  "chemistry_property_Mahler", "contact_potential_Maiorov",
  "protein_blocks_Rogov", "sequence_alignment_Dayhoff",
  "structure_alignments_Mirny"))
}
\arguments{
\item{groupingScheme}{A character vecto of length 1, determining the character
symbols used to represent amino acids grouped by their physical and chemical
properties. The available \code{groupingScheme} are "no", classic", "charge", 
"chemistry", and "hydrophobicity".}
}
\value{
A named character vector of character symbols
}
\description{
Get character symbols for grouped amino acids
}
\examples{

}
\author{
Jianhong Ou, Haibo Liu
}
\keyword{internal}
