% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proteome.example}
\alias{proteome.example}
\title{An object of \code{Proteome} representing the E. coli proteome.}
\format{A object of Proteome class}
\source{
\url{http://www.uniprot.org/}
}
\usage{
proteome.example
}
\description{
A dataset containing the E. coli proteome. The variables are as follows:
}
\details{
\itemize{
  \item proteome. A data frame containg ENTREZ_GENE IDs, protein SEQUENCES, 
                  and UniProtKB IDs and length of the sequences
  \item type. The data source: Uniprot
  \item species. Drosophila melanogaster
}
}
\keyword{datasets}
