% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seq.example}
\alias{seq.example}
\title{An object of \code{\link{dagPeptides}} representing acetylated lysine-containing
peptides.}
\format{An object of \code{\link{dagPeptides}} Class}
\usage{
seq.example
}
\description{
A dataset containing the acetylated lysine-containing peptides from Drosophila 
melanogaster. The variables are as follows:
}
\details{
\itemize{
  \item data. A data frame containing entrezgene IDs, anchorAA, anchorPos,
              peptide, anchor, upstreamOffset and downstreamOffset.
  \item peptides. A matrix of individual symbols of aligned amino acids.
  \item upstreamOffset. 
  \item downstreamOffset.
  \item type. The ID type
}
}
\keyword{datasets}
