% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDAU.R
\name{testDAU}
\alias{testDAU}
\title{Conduct differential usage test of amino acids.}
\usage{
testDAU(dagPeptides, dagBackground, groupingScheme = c("no", "classic",
  "charge", "chemistry", "hydrophobicity"), bgNoise = NA)
}
\arguments{
\item{dagPeptides}{An object of Class \code{\link{dagPeptides}}.}

\item{dagBackground}{An object of Class \code{\link{dagBackground}}.}

\item{groupingScheme}{A character vector of length 1. Available choices are
"no", "classic", "charge", "chemistry",and "hydrophobicity". It is used
to group amino acids into groups.}

\item{bgNoise}{A numeric vector of length 1. It should be in the interval of
(0, 1).}

\item{testType}{A character vector of length 1. The available options are
"fisher" and "z-test", that is "Fisher's exact test" and "Z-test". When the 
the number of sequences in the background set is too samll to perform non-
replacement subsamplint, "Fisher's exact test" is suggested.}
}
\value{
An object of Class \code{\link{testDAUresults}}.
}
\description{
Test differential usage of amino acids with or without grouping in
experimental sets and background sets.
}
\examples{
dat <- unlist(read.delim(system.file(
                                   "extdata", "grB.txt", package = "dagLogo"),
                         header = F, as.is = TRUE))
head(dat)
##prepare proteome from a fasta file
proteome <- prepareProteome(fasta = system.file("extdata",
                                                "HUMAN.fasta",
                                                package = "dagLogo"), 
                            species = "Homo sapiens")
##prepare object of dagPeptides
##prepare an object of dagPeptides
seq <- formatSequence(seq = dat, proteome = proteome, upstreamOffset = 14,
                     downstreamOffset = 15)
bg_fisher <- buildBackgroundModel(seq, background = "wholeProteome", 
                                  proteome = proteome, testType = "fisher")
bg_ztest <- buildBackgroundModel(seq, background = "wholeProteome",
                                   proteome = proteome, testType = "ztest")
 
t0 <- testDAU(seq, dagBackground = bg_ztest)
## grouded classically: nonpolar_aliphatic = c("A", "G", "L", "M", "I", "V"), 
## polar_uncharged = c("C", "P", "Q", "S", "T"), aromatic = c("F", "W", "Y"), 
## positively_charged = c("H", "K", "N", "R"), negatively_charged = c("D", "E").
t1 <- testDAU(dagPeptides = seq, dagBackground = bg_ztest, 
             groupingScheme = "classic")
## grouded on the basis of charge: positive = c("H", "K", "R"), 
## neutral = c("A", "C", "F", "G", "I", "L", "M", "N", "P", "Q", "S", "T", 
## "V","W","Y"), negative = c("D", "E")
t2 <- testDAU(dagPeptides = seq, dagBackground = bg_ztest, 
              groupingScheme = "charge")
## grouped on the basis of their chemical property: hydrophobic = c("A", "F",
## "I", "L", "M", "P", "V", "W"), polar = c("C", "G", "S", "T", "Y"), 
## basic = c("H", "K", "R"), neutral = c("N", "Q"), acidic = c("D", "E")
t3 <- testDAU(dagPeptides = seq, dagBackground = bg_ztest, 
              groupingScheme = "chemistry")
## grouped on the basis of hydrophobicity: hydrophilic = c("D", "E", "K", "N", 
## "Q", "R"), neutral = c("A", "G", "H", "P", "S", "T"), hydrophobic = c("C", 
## "F", "I", "L", "M", "V", "W", "Y")
t4 <- testDAU(dagPeptides = seq, dagBackground = bg_ztest, 
              groupingScheme = "hydrophobicity")                                   
}
\author{
Jianhong Ou, Haibo Liu
}
