% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagLogo.R
\name{colorsets}
\alias{colorsets}
\title{retrieve color setting for logo}
\usage{
colorsets(colorScheme = c("null", "classic", "charge", "chemistry",
  "hydrophobicity"))
}
\arguments{
\item{colorScheme}{could be 'null', 'charge', 'chemistry', 'classic' or 'hydrophobicity'}
}
\value{
A character vector of color scheme
}
\description{
retrieve prepared color setting for logo
}
\examples{
col <- colorsets("hydrophobicity")
}
\author{
Jianhong Ou
}
\keyword{figure}
