% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagLogo.R
\name{colorsets}
\alias{colorsets}
\title{Color sets for visualization.}
\usage{
colorsets(colorScheme = c("no", "classic", "charge", "chemistry",
  "hydrophobicity", "BLOSM50_L1", "BLOSM50_L2", "BLOSM50_L3", "BLOSM50_L4",
  "BLOSM50_L5", "BLOSM50_L6", "BLOSM50_L7", "BLOSM50_L8",
  "chemistry_property_Mahler", "contact_potential_Maiorov",
  "protein_blocks_Rogov", "sequence_alignment_Dayhoff",
  "structure_alignments_Mirny", "custom"))
}
\arguments{
\item{colorScheme}{A character vecto of length 1, determining the color scheme
based on amino acid classification methods. The available \code{colorScheme} are "no",
"classic", "charge", "chemistry", "hydrophobicity", "BLOSM50_L1", "BLOSM50_L2", 
"BLOSM50_L3", "BLOSM50_L4", "BLOSM50_L5", "BLOSM50_L6", "BLOSM50_L7", 
"BLOSM50_L8", "chemistry_property_Mahler", "contact_potential_Maiorov", 
"protein_blocks_Rogov", "sequence_alignment_Dayhoff", "structure_alignments_Mirny",
and "custom". If "custom" is used, users must define a grouping scheme using a
a list containing sublist named as "color", "symbol" and "group" using the 
function addGroupingScheme.}
}
\value{
A named character vector of colors
}
\description{
Create color encoding for visualization of a peptide sequence logo.
}
\examples{
colorsets("classic")
}
\seealso{
{\code{\link{addGroupingScheme}}}
}
\author{
Jianhong Ou, Haibo Liu
}
\keyword{internal}
