% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclass.R
\docType{class}
\name{dagBackground-class}
\alias{dagBackground-class}
\title{Class \code{"dagBackground"}}
\description{
An object of class \code{"dagBackground"} represents background model.
}
\section{Slots}{

\describe{
\item{\code{background}}{Object of class \code{"list"} records the background model}

\item{\code{permutationSize}}{code{"integer"} permutation size of background}
}}

\section{Objects from the Class}{

  Objects can be created by calls of the form 
  \code{new("dagBackground", background, permutationSize)}.
}

\keyword{classes}
