% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclass.R
\docType{class}
\name{dagBackground-class}
\alias{dagBackground-class}
\title{Class \code{\link{dagBackground}}.}
\description{
An S4 class to represent a background of a formatted, aligned peptides for 
dagLogo analysis.
}
\section{Slots}{

\describe{
\item{\code{background}}{A list of data frame. Within each n-by-1 dataframe is a the 
aligned peptides of same length.}

\item{\code{numSubsamples}}{An integer. That is the length of the \code{background} list}
}}

\author{
Jianhong Ou, Haibo Liu
}
