% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagLogo.R
\name{dagHeatmap}
\alias{dagHeatmap}
\title{plot heatmap for test results}
\usage{
dagHeatmap(testDAUresults, type = c("diff", "zscore"), ...)
}
\arguments{
\item{testDAUresults}{output of \code{\link{testDAU}}, should be an object of testDAUresults}

\item{type}{"diff" or "zscore"}

\item{\dots}{parameter could be passed to pheatmap}
}
\value{
none
}
\description{
plot heatmap for test results
}
\examples{
  data("seq.example")
  data("proteome.example")
  bg <- buildBackgroundModel(seq.example, proteome=proteome.example, permutationSize=10)
  t <- testDAU(seq.example, bg)
  dagHeatmap(t)
}
\author{
Jianhong Ou
}
\keyword{figure}
