% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBackgroundModel.R
\name{initiateBackgroundModel}
\alias{initiateBackgroundModel}
\title{Create a new \code{\link{dagBackground-class}} object for \code{\link{testDAU}}.}
\usage{
initiateBackgroundModel(background, numSubsamples = 1L, testType)
}
\arguments{
\item{background}{A list, each element of which is a vector of aligned peptide
sequences of the same length.}

\item{numSubsamples}{An integer, the number of random samplings to get 
background sequence set.}

\item{testType}{An character, the type of statistic testing : "ztest" or "fisher".}
}
\value{
An object of \code{\link{dagBackground-class}}.
}
\description{
Create a new \code{\link{dagBackground-class}} object for \code{\link{testDAU}}.
}
\author{
Haibo Liu
}
\keyword{internal}
