% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagLogo.R
\name{nameHash}
\alias{nameHash}
\title{convert group name to a single character}
\usage{
nameHash(nameScheme = c("classic", "charge", "chemistry", "hydrophobicity"))
}
\arguments{
\item{nameScheme}{could be "classic", "charge", "chemistry", "hydrophobicity"}
}
\value{
A character vector of name scheme
}
\description{
convert group name to a single character to shown in a logo
}
\author{
Jianhong Ou
}
\keyword{figure}
