% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareProteome.R
\name{prepareProteome}
\alias{prepareProteome}
\title{Create an object of Proteome class.}
\usage{
prepareProteome(source = NULL, species = NULL, destDir = getwd(),
  fastaFile, ...)
}
\arguments{
\item{source}{A character vector of length 1 or NULL. A database source from 
where the proteome sequences are to downloaded. By default, currently it is
"Uniprot". If it is NULL, then \code{fastaFile} has to be specified. The 
priority of \code{source} is higher than \code{fastaFile}.}

\item{species}{A character vector of length 1. The Latin name of a species 
confirming to the Linnaean taxonomy nomenclature system.}

\item{destDir}{A character vector of length 1. A destination directory with 
writing permission for saving downloaded sequences.}

\item{fastaFile}{A character vector of length 1. A fasta file name from which
to read in protein sequences.}

\item{...}{other parameters passing to \link{download.file}.}
}
\value{
An object of Proteome
}
\description{
Create an object of Proteome class by downloading a whole proteome data from 
Uniprot for a given organism of an NCBI taxonomy ID or by using peptide sequences
in a fasta file.
}
\examples{
## Prepare an objecto of Proteome from the UniProt database
if(interactive()) 
{
    proteome <- prepareProteome(source = "UniProt", species = "Homo sapiens")
}

## Prepare n objecto of Proteome from a fasta file
fasta <- system.file("extdata", "Drosophila.melanogaster.fa", package="dagLogo")
proteome <- prepareProteome(source = NULL, species = "Drosophila melanogaster", fastaFile=fasta)
}
