% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclass.R
\docType{class}
\name{Proteome-class}
\alias{Proteome-class}
\title{Class \code{\link{Proteome}}.}
\description{
An S4 class to represent a whole proteome for dagLogo analysis.
}
\section{Slots}{

\describe{
\item{\code{proteome}}{A data frame.}

\item{\code{type}}{A character vector of length 1. Available options :"UniProt", and 
"fasta".}

\item{\code{species}}{A character vector of length 1, a conventional Latin name for 
a species.}
}}

\author{
Jianhong Ou
}
